+++ packages/vue/src/install.ts
@@ -8,7 +8,7 @@ export default function install(_Vue: typeof Vue, config: CuriConfig): void {
   let curi: CuriPluginObject = { config, response: null, action: null };
   _Vue.use(CuriPlugin, { curi });
 
-  config.subscribe((response: Response, action: Action): void => {
+  config.respond((response: Response, action: Action): void => {
     curi.response = response;
     curi.action = action;
   });